+++ backend/app/routers/users.py
@@ -1,42 +1,54 @@
-from fastapi import APIRouter, Depends, status
+from fastapi import APIRouter, Depends, status, HTTPException
 from sqlalchemy.orm import Session
 from typing import List
+
 from ..database import get_db
 from ..schemas.user import UserCreate, UserResponse
 from ..services.user_services import UserService
+from backend.app.dependencies.auth import get_current_user  # dependency для текущего пользователя
 
 router = APIRouter(
     prefix="/users",
     tags=["users"]
 )
 
+
 @router.get("/", response_model=List[UserResponse], status_code=status.HTTP_200_OK)
-def get_all_users(db: Session = Depends(get_db)):
+def get_all_users(
+    current_user=Depends(get_current_user),
+    db: Session = Depends(get_db)
+):
     user_service = UserService(db)
     return user_service.get_all_users()
 
 @router.get("/{user_id}", response_model=UserResponse, status_code=status.HTTP_200_OK)
-def get_user_by_id(user_id: int, db: Session = Depends(get_db)):
+def get_user_by_id(user_id: int, current_user = Depends(get_current_user), db: Session = Depends(get_db)):
     user_service = UserService(db)
     return user_service.get_user_by_id(user_id)
 
 @router.get("/nickname/{nickname}", response_model=UserResponse, status_code=status.HTTP_200_OK)
-def get_user_by_nickname(nickname: str, db: Session = Depends(get_db)):
+def get_user_by_nickname(nickname: str, current_user=Depends(get_current_user), db: Session = Depends(get_db)):
     user_service = UserService(db)
     return user_service.get_user_by_nickname(nickname)
+
 
 @router.post("/", response_model=UserResponse, status_code=status.HTTP_201_CREATED)
 def create_user(user: UserCreate, db: Session = Depends(get_db)):
     user_service = UserService(db)
     return user_service.create_user(user)
 
+
 @router.put("/{user_id}", response_model=UserResponse, status_code=status.HTTP_200_OK)
-def update_user(user_id: int, user: UserCreate, db: Session = Depends(get_db)):
+def update_user(user_id: int, user: UserCreate, current_user=Depends(get_current_user), db: Session = Depends(get_db)):
+    if user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this user")
     user_service = UserService(db)
     return user_service.update_user(user_id, user)
 
 @router.delete("/{user_id}", status_code=status.HTTP_204_NO_CONTENT)
-def delete_user(user_id: int, db: Session = Depends(get_db)):   
+def delete_user(user_id: int, current_user=Depends(get_current_user), db: Session = Depends(get_db)):
+    if user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to delete this user")
     user_service = UserService(db)
     user_service.delete_user(user_id)
     return None
